Crate sgp4

Simplified Perturbations Models (SGP4)

The Simplified Perturbations Models are a set of models used for satellites and objects relative to the Earth-centered inertial coordinate system. These are often referred to collectively as SGP4 because of how popular that particular code is and how it's used with nearly all low Earth orbit satellites.

The SGP4 and SDP4 models were published as FORTRAN IV in 1988. It has also been ported to C. This is a port to Rust.

Original paper: Hoots_Roehrich_1980_SPACETRACK_REPORT_NO_3.pdf

Modules

coordinates

tle

Constants

A30
$A_{3,0} = -J_3a_E^3$

J3
$J_3 = -2.53881 \times 10^{-4}$: the third gravitational zonal harmonic of the Earth

QS4
qs4 (?)

RE
$R_\oplus = 1.0$ Radius of the Earth (in Earth Radii).

S
S (?)

XKMPER
$6378.135$ kilometers/Earth radii.

k2
$k_2 = 5.413080 \times 10^{-4}$ Harmonic gravity constant for the SGP4 model. Defined as $\frac{1}{2}J_2aE^2$.

ke
$k_e = 7.43669161 \times 10^{-2}$ Orbital constant for Earth defined as $\sqrt{GM_{\oplus}}$ where $G$ is Newton’s universal gravitational constant and $M_{\oplus}$ is the mass of the Earth. Units: $(\frac{\mathrm{Earth\ radii}}{\mathrm{minute}})^{\frac{3}{2}}$

Functions

propagate