Struct sgp4::tle::TLE

pub struct TLE {
    pub name: String,
    pub sat_number: u32,
    pub classification: char,
    pub int_designator: String,
    pub epoch_year: u16,
    pub epoch_day: f64,
    pub first_mean_motion: f64,
    pub second_mean_motion: f64,
    pub bstar: f64,
    pub tle_version: u16,
    pub i: f64,
    pub raan: f64,
    pub e: f64,
    pub omega: f64,
    pub mean_anomaly: f64,
    pub mean_motion: f64,
    pub revolution_number: u32,
}

Satellite elements from a TLE file

Read "Two Line Element" files that are a standard way of distributing defined orbits.

Fields

name

Name of the object

sat_number

The Satellite Catalog Number, a sequential 5-digit number assigned by United States Space Command to all Earth orbiting satellites.

classification

Classification (U=Unclassified)

int_designator

International Designator

epoch_year

Epoch Year

epoch_day

Epoch Day

first_mean_motion

First Time Derivative of the Mean Motion

second_mean_motion

Second Time Derivative of Mean Motion

bstar

BSTAR drag term

tle_version

Element set number (Incremented when a new TLE is generated for this object)

i

Inclination

raan

Right ascension of the ascending node

e

Eccentricity

omega

Argument of perigee

mean_anomaly

Mean Anomaly

mean_motion

Mean Motion (revolutions per day)

revolution_number

Revolution number at epoch (revolutions)